Reconfigured the way the program is organized
If we're going to start using angular to execute functions from the
front end then we're going to need to be pretty organized. I just
reconfigured where everything is. Made appropriate changes to app.js so
everything still works.

Turns out angular as a frontend language doesn't play nice with node.js,
so we're going to have to start using HTTP requests from the client to
the server most likely. Will explain more in person.